Abstracting out the common flags code between Symbol and Modifiers into HasFlags.  This patch includes only the non-invasive changes: the HasFlags trait is not integrated into those classes.  The remainder is complete but I'm checking this part in case anyone would like to throw some feedback my way at this point: general comments, or the open issues noted in comments in HasFlags.scala.
This commit also eliminates the (now unused) MONOMORPHIC flag and
includes utility methods for reflective analysis of a Flags object
which generate code and comments based on the accessors found.
See the comment at the top of symtab.Flags and the flagToString
implementation in that class for illustration.

I haven't tested the very latest, but a slightly older incarnation
without material differences showed no measurable performance change.

No specific review, but comments welcome.

git-svn-id: https://lampsvn.epfl.ch/svn-repos/scala/scala/trunk@23226 5e8d7ff9-d8ef-0310-90f0-a4852d11357a