do not add entry in InnerClass attribute if outer name is null (according to comment for outerName method) in order to remove the following warning message reported by the dx tool (Android):
warning: Ignoring InnerClasses attribute for an anonymous inner class
(test$$anonfun$1) that doesn't come with an
associated EnclosingMethod attribute. This class was probably produced by a
compiler that did not target the modern .class file format. The recommended
solution is to recompile the class from source, using an up-to-date compiler
and without specifying any "-target" type options. The consequence of ignoring
this warning is that reflective operations on this class will incorrectly
indicate that it is *not* an inner class.



git-svn-id: https://lampsvn.epfl.ch/svn-repos/scala/scala/trunk@23785 5e8d7ff9-d8ef-0310-90f0-a4852d11357a