Removed parens from calls to Name#toString.
This commit doesn't include the line which would change anything
interesting (that being making toString in Name a val), it is all
the changes elsewhere which one would have to make to do that. I
didn't want to waste them. A quirk of scala is that if you call an
empty-parameter-list method:

  foo.bar() // e.g. x.toString()
  foo.bar   // where you could have called x.toString

then if you decide in the future to make it a val, every call site must
potentially be modified. This is a bit of a bummer. One can try to
defend against that future decision by eschewing parentheses, but it's
not very satisfying or effective. No review.

git-svn-id: https://lampsvn.epfl.ch/svn-repos/scala/scala/trunk@25659 5e8d7ff9-d8ef-0310-90f0-a4852d11357a