Fixing isTrivial in NullaryMethod.
1) Made NullaryMethodType a SimpleTypeProxy. (This is incidental
to isTrivial, it just seemed like the right thing to do.)

2) Modified implementations of isTrivial for MethodType and
NullaryMethodType to only be considered trivial if there are no
annotations on the return type.  This means asSeenFrom will not
be skipped and the continuations plugin will work properly.
There are a lot of other angles one could take, so review by moors, rompf.

git-svn-id: https://lampsvn.epfl.ch/svn-repos/scala/scala/trunk@25710 5e8d7ff9-d8ef-0310-90f0-a4852d11357a