Fix for multiple evaluation in structural calls.
An interesting bug during cleanup: runtime checks on the target of a
structural invocation duplicated the selection without regard for the
fact that it might be an expression. So if the name of the method being
invoked allowed the possibility that the target was a primitive type
(such as "toInt") the expression would be evaluated three times.

Closes SI-5080, no review.

git-svn-id: https://lampsvn.epfl.ch/svn-repos/scala/scala/trunk@25838 5e8d7ff9-d8ef-0310-90f0-a4852d11357a